zsh: move EDITOR to zshenv
Setting EDITOR late in .zshrc means that its not set when ZLE is
initialized. ZLE uses EDITOR to decide whether use the `emacs` keymap or
`viins` keymap, but which is not respected when set where it currently
is. By moving it to .zshenv, ZLE will see it and correctly infer `viins`
by default. This also helps with consistency between terminals (some
terminals embedded in other applications would "see" `EDITOR` and start
with the `viins` keymap while others wouldn't see `EDITOR` and start
with the `emacs` keymap).